requirements.txt
apiclient==1.0.4
asn1crypto==1.1.0
astroid==2.3.1
awesome-slugify==1.4
backports.csv==1.0.7
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-allauth==0.25.2
django-autofixture==0.12.1
django-braces==1.9.0
django-cors-headers==2.3.0
django-crispy-forms==1.4.0
django-extensions==2.2.3
django-filter==2.2.0
django-hashedfilenamestorage==1.0
django-import-export==1.2.0
django-model-utils==1.4.0
django-nested-admin==3.2.3
django-nose==1.4.6
django-pipeline==1.3.11
django-pure-pagination==0.3.0
django-rest-swagger==2.2.0
django-reversion==1.9.3
django-widget-tweaks==1.3
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
easy-thumbnails==2.6
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==2.0.3
flake8==3.7.8
futures==3.1.1
gevent==1.4.0
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
greenlet==0.4.15
httplib2==0.13.1
httplib2shim==0.0.3
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jsonfield==2.0.2
lazy-object-proxy==1.4.2
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
packaging==19.2
pep8==1.5.7
Pillow==6.2.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.2
pyOpenSSL==17.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.1
rauth==0.7.3
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
text-unidecode==1.3
typed-ast==1.4.0
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
